fix(ledger): widen operational certificate counter and KES period
The operational certificate sequence number and KES period were uint32 in
the Shelley-family and Babbage-family header bodies. cardano-ledger decodes
the counter as Word64 and the period as a Word newtype with plain decCBOR
and no bound, and the CDDL declares both as uint .size 8, so a header
carrying either at or above 2^32 failed to decode instead of being rejected
by the operational certificate rules.

Fixes #2245

+++ consensus/validate.go
@@ -87,8 +87,8 @@ type ValidateHeaderInput struct {
 
 	// OpCert fields
 	OpCertHotVkey        []byte
-	OpCertSequenceNumber uint32
-	OpCertKesPeriod      uint32
+	OpCertSequenceNumber uint64
+	OpCertKesPeriod      uint64
 	OpCertSignature      []byte
 
 	// Previous header for chain validation
@@ -456,7 +456,7 @@ func (v *HeaderValidator) validateKESPeriod(input *ValidateHeaderInput) error {
 	}
 
 	currentKESPeriod := input.Slot / v.slotsPerKESPeriod
-	opCertKESPeriod := uint64(input.OpCertKesPeriod)
+	opCertKESPeriod := input.OpCertKesPeriod
 
 	// OpCert cannot be from the future
 	if currentKESPeriod < opCertKESPeriod {
@@ -510,7 +510,7 @@ func (v *HeaderValidator) validateKESSignature(
 
 	// Calculate evolution period
 	currentKESPeriod := input.Slot / v.slotsPerKESPeriod
-	opCertKESPeriod := uint64(input.OpCertKesPeriod)
+	opCertKESPeriod := input.OpCertKesPeriod
 	// Guard against underflow if OpCert KES period is in the future
 	if currentKESPeriod < opCertKESPeriod {
 		return fmt.Errorf(
@@ -569,8 +569,8 @@ func (v *HeaderValidator) validateOpCertSignature(
 	// (hot_vkey || sequence_number || kes_period), not a CBOR encoding.
 	opCertBody := common.OpCertSignableBytes(
 		input.OpCertHotVkey,
-		uint64(input.OpCertSequenceNumber),
-		uint64(input.OpCertKesPeriod),
+		input.OpCertSequenceNumber,
+		input.OpCertKesPeriod,
 	)
 
 	// Verify Ed25519 signature
